requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
gunicorn==20.1.0
google-cloud-storage==2.9.0
flask-cors
gradio_client
gradio
aiohttp